Testability
This extracts the actual DeepLApi call into its own component.
The TranslationSuggestionControllerTest is mocking the API call.
